Add Rack::Session::Pool :allow_fallback option
This fixes a theoretical security issue where a lookup of the session
ID uses a non-constant time algorithm (such as a database index), and
only for cases where there is a session that existed prior to an
rack version that added support for private session IDs.

This defaults :allow_fallback to true for backwards compatibility,
but we may want to make the default false in Rack 3.

Fixes #1431